+++ AirsoftMatchMaker.Infrastructure/Data/Entities/Weapon.cs
@@ -14,8 +14,8 @@ namespace AirsoftMatchMaker.Infrastructure.Data.Entities
         [Required]
         [MaxLength(200)]
         public string Description { get; set; } = null!;
-        public decimal FeetPerSecond { get; set; }
-        public decimal FireRate { get; set; }
+        public double FeetPerSecond { get; set; }
+        public double FireRate { get; set; }
         public decimal Price { get; set; }
         public string? ImageUrl { get; set; }
         public int AverageAmmoExpendedPerGame { get; set; }
